requirements.txt
openai
tiktoken
tqdm
python-dotenv